Light formatting changes
I tried to fix as many of the visible formatting errors as I could.  The
underlying issues with formatting remain and will likely require a full
redesign of the document however.  I also removed the Hover Truck entry
as it is a duplicate of the one found in the core rulebook.