Force Organization Chart added
I added the force organization chart in a new section.  Also tried to
correct the formatting on armor value zaku diagram by converting it to a
jpeg screenshot taken from the PDF.  I think that is the route we'll
have to go with all of the to keep them from spazzing out every time
anything is changed.