Special rules, 40k references and gun changes
I made several additions to the core rules.
-I added a universal special rules section for the rules already
outlined in the core rules and 0080.  I also changed the references to
the the Warhammer 40k rule book to the Core rules
-I added the 90mm bulpup machine gun to the GM as it is listed as an
option on Gundam wiki
-I added the MMP-70C to the core rules as it was listed as an option for
several Zeon suits but was not listed in the armory.
-I added a description for independent character to the special rules
page.  The way the rule is written in 40k, dealing primarily with
squads, was not particularly useful for the game as it stands.  I
changed it to incorporate rules from the voxcaster from Codex Imperial
Guard.  It lets any suit within 8" of a character.  I also removed it
from all non-HQ characters to balance it a bit.
-I also added a few terms to the glossary